move stickers directory
