+++ assets/src/edit-story/elements/sticker/display.js
@@ -17,7 +17,7 @@
  * Internal dependencies
  */
 import StoryPropTypes from '../../types';
-import stickers from './stickers';
+import stickers from '../../stickers';
 
 const style = {
   display: 'block',
