+++ assets/src/edit-story/stickers/index.js
