+++ assets/src/edit-story/stickers/sampleSticker.js
